# docs/api/swagger-fr.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: 'Fake ID Generator - Middlware API'
  description: "Fake ID Generator est un générateur de cartes d'identité fictives qui vise à offrir une expérience pratique et utile pour diverses applications, notamment le développement de jeux de rôle et d'autres projets nécessitant des données d'identité réalistes."
host: 'api.fake-id-generator.com'
basePath: '/v1'
schemes:
  - 'https'
  - 'http'
paths:
...
  /identity/simple:
    get:
      tags:
        - Identity
      summary: 'Génère un profil complet avec toutes les informations'
      operationId: 'generateFullProfile'
      produces:
        - 'application/json'
      parameters:
        - name: 'firstname'
          in: 'query'
          description: "Prénom utilisé pour générer le genre et d'autres détails"
          required: false
          type: 'string'
        - name: 'gender'
          in: 'query'
          description: 'Genre utilisé pour influencer la génération du nom et du visage'
          required: false
          type: 'string'
        - name: 'country'
          in: 'query'
          description: "Pays utilisé pour influencer la génération du lieu de naissance et de l'âge"
          required: false
          type: 'string'
      responses:
        200:
          description: 'Profil complet généré avec succès'
          schema:
            type: 'object'
            properties:
              id:
                type: 'string'
                example: 'uuid-v4-string'
              firstName:
                type: 'string'
                example: 'Jean'
              lastName:
                type: 'string'
                example: 'Dupont'
              gender:
                type: 'string'
                example: 'male'
              nationality:
                type: 'string'
                example: 'French'
              birthplace:
                type: 'string'
                example: 'Paris'
              age:
                type: 'integer'
                example: 29
              faceImage:
                type: 'string'
                example: '/9j/4AAQSkZJRgABAQEAAAAAAAD...'
  /identity/createdWithName:
    get:
      tags:
        - Identity
      summary: 'Génère un profil complet basé sur un prénom donné'
      operationId: 'generateFullProfileByName'
      produces:
        - 'application/json'
      parameters:
        - name: 'firstname'
          in: 'query'
          description: 'Prénom utilisé pour générer le profil complet'
          required: true
          type: 'string'
      responses:
        200:
          description: 'Profil complet généré basé sur le prénom fourni'
          schema:
            type: 'object'
            properties:
              id:
                type: 'string'
                example: 'uuid-v4-string'
              firstName:
                type: 'string'
                example: 'Jean'
              lastName:
                type: 'string'
                example: 'Dupont'
              gender:
                type: 'string'
                example: 'male'
              nationality:
                type: 'string'
                example: 'French'
              birthplace:
                type: 'string'
                example: 'Paris'
              age:
                type: 'integer'
                example: 29
              faceImage:
                type: 'string'
                example: '/9j/4AAQSkZJRgABAQEAAAAAAAD...'
  /identity/createdWithNationality:
    get:
      tags:
          - Identity
      summary: 'Génère un profil complet basé sur une nationalité donnée'
      operationId: 'generateFullProfileByNationality'
      produces:
        - 'application/json'
      parameters:
        - name: 'nationality'
          in: 'query'
          description: 'Nationalité utilisée pour générer le profil complet'
          required: true
          type: 'string'
      responses:
        200:
          description: 'Profil complet généré basé sur la nationalité fournie'
          schema:
            type: 'object'
            properties:
              id:
                type: 'string'
                example: 'uuid-v4-string'
              firstName:
                type: 'string'
                example: 'Jean'
              lastName:
                type: 'string'
                example: 'Dupont'
              gender:
                type: 'string'
                example: 'male'
              nationality:
                type: 'string'
                example: 'French'
              birthplace:
                type: 'string'
                example: 'Paris'
              age:
                type: 'integer'
                example: 29
              faceImage:
                type: 'string'
                example: '/9j/4AAQSkZJRgABAQEAAAAAAAD...'
...
securityDefinitions:
  ApiKeyAuth:
    type: 'apiKey'
    in: 'header'
    name: 'Authorization'
security:
  - ApiKeyAuth: []